requirements.txt
pydantic
python-dotenv
numpy
scipy
black
openai
anthropic
google-generativeai
streamlit
pandas
altair
charset_normalizer
watchdog
pymongo
numpy
scipy
trueskill
groq
